requirements.txt
appdirs==1.4.4
asgiref==3.2.10
distlib==0.3.1
Django==3.1
django-ckeditor==6.0.0
django-js-asset==1.2.2
filelock==3.0.12
Pillow==7.2.0
psycopg2==2.8.5
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
virtualenv==20.0.31
